refactor: api와 단축 url 접속 로직에 관한 컨트롤러를 따로 분리
